Refactored loader ahead of starting event loader
