+++ collex-loader/.gitignore
@@ -1,2 +1,3 @@
 tmp
 *~
+__pycache__
