+++ collex-loader/config/app-config.json
@@ -1,5 +1,5 @@
 {
-    "inputFile": "data/CE_Details_for_loading_into_RM.csv",
+    "inputFiles": ["data/CE_Details_for_loading_into_RM.csv"],
     "dryRun": false,
     "api":{
         "url": "http://localhost:8145/collectionexercises",
