+++ collex-loader/load.py
@@ -1,27 +1,17 @@
 #!/usr/bin/python
 import argparse
-import csv
 from functools import partial
 import json
-
 import requests
-
+from file_processing import process_files
 
 def parse_args():
     parser = argparse.ArgumentParser(description='Load collection exercise CSV.')
     parser.add_argument("config", help="Configuration file")
     return parser.parse_args() 
 
-
-def map_columns(column_mappings, row):
-    new_row = dict()
-    for key, value in row.items():
-        if key and value:
-            new_row[column_mappings[key] if column_mappings[key] else key] = value
-    return new_row
-
-
 def post_collex(data, url, user, password):
+    data = clean_row(data)
     response = requests.post(url, json=data, auth=(user, password))
 
     status_code = response.status_code
@@ -31,6 +21,7 @@ def post_collex(data, url, user, password):
 
 
 def dump_collex(data):
+    data = clean_row(data)
     survey_id = data['surveyRef']
     period = data['exerciseRef']
     if survey_id and period:
@@ -45,22 +36,12 @@ def clean_row(row):
     return row
         
 
-def process_file(filename, row_handler):
-    with open(filename) as fp:
-        reader = csv.DictReader(fp)
-        for row in reader:
-            new_row = map_columns(column_mappings, row)
-
-            if new_row:
-                row_handler(data=clean_row(new_row))
-
-
 if __name__ == '__main__':
     args = parse_args()
     print("Config filename: %s" % args.config)
     config = json.load(open(args.config))
-    input_file = config['inputFile']
-    print("Input filename: %s" % input_file)
+    input_files = config['inputFiles']
+    print("Input filenames: %s" % input_files)
     column_mappings = config['columnMappings']
     api_url = config['api']['url']
     print("API URL: %s" % api_url)
@@ -69,4 +50,4 @@ if __name__ == '__main__':
                                                             user=config['api']['user'],
                                                             password=config['api']['password'])
 
-    process_file(input_file, row_handler)
+    process_files(input_files, row_handler, column_mappings)
